requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
distro==1.9.0
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
groq==0.11.0
gspread==6.1.4
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-core==0.3.15
langchain-text-splitters==0.3.2
langsmith==0.1.142
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.0
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
propcache==0.2.0
prov==2.0.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
serpapi==0.1.5
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
tenacity==9.0.0
tokenizers==0.20.3
tqdm==4.67.0
traits==6.4.3
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.2
yarl==1.17.1
